+++ src/core/transaction.cpp
@@ -1,5 +1,1 @@
-//
-// Created by maks on 1/6/24.
-//
-
 #include "transaction.h"