+++ src/core/utils.h
@@ -1,0 +1,8 @@
+#ifndef FCOIN_UTILS_H
+#define FCOIN_UTILS_H
+
+#include "constants.h"
+
+bool startsWith(const str& input, const str& prefix);
+
+#endif //FCOIN_UTILS_H